beautify login and singup
