+++ public/components/Profile/profile.css
@@ -20,8 +20,8 @@ header#app{
     border: 2px solid #2dc68b;
     margin-top: 10px;
     border-radius: 50%;
-    height: 280px;
-    width: 280px;
+    height: 200px;
+    width: 200px;
     object-position: center center;
     object-fit: fill;
     box-sizing: border-box;
