+++ public/main.css
@@ -1,5 +1,5 @@
 @import url('./components/Profile/profile.css');
-@import url('./components/Authorization/auth.css');
+@import url('components/Signup/signup.css');
 @import url('./components/MainPage/mainPage.css');
